+++ member/src/main/java/com/mossy/boundedContext/app/seller/SellerRequestAdminFacade.java
@@ -7,9 +7,9 @@ import com.mossy.boundedContext.domain.seller.SellerRequest;
 import com.mossy.boundedContext.domain.user.User;
 import com.mossy.exception.DomainException;
 import com.mossy.exception.ErrorCode;
-import com.mossy.boundedContext.out.seller.SellerRepository;
-import com.mossy.boundedContext.out.user.RoleRepository;
-import com.mossy.boundedContext.out.user.UserRepository;
+import com.mossy.boundedContext.out.repository.seller.SellerRepository;
+import com.mossy.boundedContext.out.repository.user.RoleRepository;
+import com.mossy.boundedContext.out.repository.user.UserRepository;
 import com.mossy.global.eventPublisher.EventPublisher;
 import com.mossy.shared.member.domain.enums.SellerRequestStatus;
 import com.mossy.shared.member.domain.role.Role;
