+++ member/src/main/java/com/mossy/boundedContext/app/seller/SellerRequestUserFacade.java
@@ -5,8 +5,8 @@ import com.mossy.boundedContext.domain.user.User;
 import com.mossy.exception.DomainException;
 import com.mossy.exception.ErrorCode;
 import com.mossy.boundedContext.in.dto.request.SellerRequestCreateRequest;
-import com.mossy.boundedContext.out.seller.SellerRequestRepository;
-import com.mossy.boundedContext.out.user.UserRepository;
+import com.mossy.boundedContext.out.repository.seller.SellerRequestRepository;
+import com.mossy.boundedContext.out.repository.user.UserRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
