+++ member/src/main/java/com/mossy/boundedContext/app/user/SignupUseCase.java
@@ -5,8 +5,8 @@ import com.mossy.exception.DomainException;
 import com.mossy.exception.ErrorCode;
 import com.mossy.boundedContext.global.ut.EncryptionUtils;
 import com.mossy.boundedContext.in.dto.request.SignupRequest;
-import com.mossy.boundedContext.out.user.RoleRepository;
-import com.mossy.boundedContext.out.user.UserRepository;
+import com.mossy.boundedContext.out.repository.user.RoleRepository;
+import com.mossy.boundedContext.out.repository.user.UserRepository;
 import com.mossy.shared.member.domain.enums.UserStatus;
 import com.mossy.shared.member.domain.role.Role;
 import com.mossy.shared.member.domain.role.RoleCode;
